Renaming module for naming conflicts
